chore: incorporate comments
