+++ apps/tangle-dapp/app/providers.tsx
@@ -11,21 +11,24 @@ import NextAdapterApp from 'next-query-params/app';
 import qs from 'query-string';
 import { type PropsWithChildren, type ReactNode } from 'react';
 import { QueryParamProvider } from 'use-query-params';
+import z from 'zod';
 
 const appEvent = new AppEvent();
 
-const Providers = ({ children }: PropsWithChildren): ReactNode => {
-  const blockedRegions =
-    typeof process.env['OFAC_REGIONS'] === 'string' &&
-    process.env['OFAC_REGIONS'].length > 0
-      ? JSON.parse(process.env['OFAC_REGIONS'])
-      : undefined;
+const envSchema = z.object({
+  OFAC_REGIONS: z
+    .preprocess((val) => JSON.parse(String(val)), z.array(z.string()))
+    .optional(),
+  OFAC_COUNTRY_CODES: z
+    .preprocess((val) => JSON.parse(String(val)), z.array(z.string()))
+    .optional(),
+});
 
-  const blockedCountryCodes =
-    typeof process.env['OFAC_COUNTRY_CODES'] === 'string' &&
-    process.env['OFAC_COUNTRY_CODES'].length > 0
-      ? JSON.parse(process.env['OFAC_COUNTRY_CODES'])
-      : undefined;
+const Providers = ({ children }: PropsWithChildren): ReactNode => {
+  const {
+    OFAC_COUNTRY_CODES: blockedCountryCodes,
+    OFAC_REGIONS: blockedRegions,
+  } = envSchema.parse(process.env);
 
   return (
     <NextThemeProvider>
